Convert fixGrammarTypes to ts
This is a typescript project, so the helper script is also written in
typescript. The script is moved to the "dev" directory, since "build" is
the default output folder when running sushi.
